+++ R/radae.R
@@ -19,7 +19,12 @@
 #'
 #' @template return_data.frame
 #'
+#' @importFrom dplyr arrange case_when group_by inner_join mutate n rowwise select ungroup rename bind_cols
+#' left_join setdiff
+#' @importFrom magrittr %>%
+#' @importFrom tibble tribble
 #' @importFrom rlang .data
+#' @importFrom utils read.table
 #'
 #'
 #' @export
@@ -30,7 +35,7 @@
 #' radae(ADSL, seed = 2)
 #'
 #' # Add metadata.
-#' aag <- utils::read.table(
+#' aag <- read.table(
 #' sep = ',', header = TRUE,
 #' text = paste(
 #'   "NAMVAR,SRCVAR,GRPTYPE,REFNAME,REFTERM,SCOPE",
@@ -81,7 +86,7 @@ radae <- function(ADSL, # nolint
 
   lookup_ae <- if_null( # nolint
     lookup,
-    tibble::tribble(
+    tribble(
       ~AEBODSYS, ~AELLT,          ~AEDECOD,        ~AEHLT,        ~AEHLGT,      ~AETOXGR, ~AESOC, ~AESER, ~AEREL,
       "cl A.1",  "llt A.1.1.1.1", "dcd A.1.1.1.1", "hlt A.1.1.1", "hlgt A.1.1", "1",        "cl A", "N",    "N",
       "cl A.1",  "llt A.1.1.1.2", "dcd A.1.1.1.2", "hlt A.1.1.1", "hlgt A.1.1", "2",        "cl A", "Y",    "N",
@@ -97,7 +102,7 @@ radae <- function(ADSL, # nolint
   )
 
   AAG <- if (is.null(lookup_aag)) { # nolint
-    utils::read.table(
+    read.table(
       sep = ",", header = TRUE,
       text = paste(
         "NAMVAR,SRCVAR,GRPTYPE,REFNAME,REFTERM,SCOPE",
@@ -120,7 +125,7 @@ radae <- function(ADSL, # nolint
     function(id, sid) {
       n_aes <- sample(c(0, seq_len(max_n_aes)), 1)
       i <- sample(seq_len(nrow(lookup_ae)), n_aes, TRUE)
-      dplyr::mutate(
+      mutate(
         lookup_ae[i, ],
         USUBJID = id,
         STUDYID = sid
@@ -131,8 +136,8 @@ radae <- function(ADSL, # nolint
   ) %>%
     Reduce(rbind, .) %>%
     `[`(c(10, 11, 1, 2, 3, 4, 5, 6, 7, 8, 9)) %>%
-    dplyr::mutate(AETERM = gsub("dcd", "trm", .data$AEDECOD)) %>%
-    dplyr::mutate(AESEV = dplyr::case_when(
+    mutate(AETERM = gsub("dcd", "trm", .data$AEDECOD)) %>%
+    mutate(AESEV = case_when(
       AETOXGR == 1 ~ "MILD",
       AETOXGR %in% c(2, 3) ~ "MODERATE",
       AETOXGR %in% c(4, 5) ~ "SEVERE"
@@ -149,33 +154,33 @@ radae <- function(ADSL, # nolint
   )
 
   # merge ADSL to be able to add AE date and study day variables
-  ADAE <- dplyr::inner_join(ADSL, ADAE,by = c("STUDYID", "USUBJID")) %>% # nolint
-    dplyr::rowwise() %>%
-    dplyr::mutate(trtsdt_int = as.numeric(as.Date(.data$TRTSDTM))) %>%
-    dplyr::mutate(trtedt_int = dplyr::case_when(
+  ADAE <- inner_join(ADSL, ADAE,by = c("STUDYID", "USUBJID")) %>% # nolint
+    rowwise() %>%
+    mutate(trtsdt_int = as.numeric(as.Date(.data$TRTSDTM))) %>%
+    mutate(trtedt_int = case_when(
       !is.na(.data$TRTEDTM) ~ as.numeric(as.Date(.data$TRTEDTM)),
       is.na(.data$TRTEDTM) ~ floor(trtsdt_int + (study_duration_secs) / 86400)
     )) %>%
-    dplyr::mutate(ASTDTM = as.POSIXct(
+    mutate(ASTDTM = as.POSIXct(
       (sample(.data$trtsdt_int:.data$trtedt_int, size = 1) * 86400),
       origin = "1970-01-01")) %>%
-    dplyr::mutate(astdt_int = as.numeric(as.Date(.data$ASTDTM))) %>%
-    dplyr::mutate(ASTDY = ceiling(as.numeric(difftime(.data$ASTDTM, .data$TRTSDTM, units = "days")))) %>%
+    mutate(astdt_int = as.numeric(as.Date(.data$ASTDTM))) %>%
+    mutate(ASTDY = ceiling(as.numeric(difftime(.data$ASTDTM, .data$TRTSDTM, units = "days")))) %>%
     # add 1 to end of range incase both values passed to sample() are the same
-    dplyr::mutate(AENDTM = as.POSIXct(
+    mutate(AENDTM = as.POSIXct(
       (sample(.data$astdt_int:(.data$trtedt_int + 1), size = 1) * 86400),
       origin = "1970-01-01")) %>%
-    dplyr::mutate(AENDY = ceiling(as.numeric(difftime(.data$AENDTM, .data$TRTSDTM, units = "days")))) %>%
-    dplyr::select(-.data$trtsdt_int, -.data$trtedt_int, -.data$astdt_int) %>%
-    dplyr::ungroup() %>%
-    dplyr::arrange(.data$STUDYID, .data$USUBJID, .data$ASTDTM, .data$AETERM)
+    mutate(AENDY = ceiling(as.numeric(difftime(.data$AENDTM, .data$TRTSDTM, units = "days")))) %>%
+    select(-.data$trtsdt_int, -.data$trtedt_int, -.data$astdt_int) %>%
+    ungroup() %>%
+    arrange(.data$STUDYID, .data$USUBJID, .data$ASTDTM, .data$AETERM)
 
   ADAE <- ADAE %>% # nolint
-    dplyr::group_by(.data$USUBJID) %>%
-    dplyr::mutate(AESEQ = seq_len(dplyr::n())) %>%
-    dplyr::mutate(ASEQ = .data$AESEQ) %>%
-    dplyr::ungroup() %>%
-    dplyr::arrange(
+    group_by(.data$USUBJID) %>%
+    mutate(AESEQ = seq_len(n())) %>%
+    mutate(ASEQ = .data$AESEQ) %>%
+    ungroup() %>%
+    arrange(
       .data$STUDYID,
       .data$USUBJID,
       .data$ASTDTM,
@@ -204,24 +209,24 @@ radae <- function(ADSL, # nolint
   )
 
   ADAE <- ADAE %>% # nolint
-    dplyr::mutate(AEOUT = factor(ifelse(
+    mutate(AEOUT = factor(ifelse(
       .data$AETOXGR == "5",
       "FATAL",
       as.character(sample_fct(outcomes, nrow(ADAE), prob = c(0.1, 0.2, 0.1, 0.3, 0.3)))
     ))) %>%
-    dplyr::mutate(AEACN = factor(ifelse(
+    mutate(AEACN = factor(ifelse(
       .data$AETOXGR == "5",
       "NOT EVALUABLE",
       as.character(sample_fct(actions, nrow(ADAE), prob = c(0.05, 0.05, 0.05, 0.01, 0.05, 0.1, 0.45, 0.1, 0.05)))
     ))) %>%
-    dplyr::mutate(AESDTH = dplyr::case_when(
+    mutate(AESDTH = case_when(
       .data$AEOUT == "FATAL" ~ "Y",
       TRUE ~ "N"
     )) %>%
-    dplyr::mutate(TRTEMFL = ifelse(.data$ASTDTM >= .data$TRTSDTM, "Y", "")) %>%
-    dplyr::mutate(
-      ANL01FL = ifelse(.data$TRTEMFL == "Y" & .data$ASTDTM <= .data$TRTEDTM + 2592000, "Y", "")
-    )
+    mutate(TRTEMFL = ifelse(.data$ASTDTM >= .data$TRTSDTM, "Y", "")) %>%
+    mutate(AECONTRT = sample(c("Y", "N"), prob = c(0.4, 0.6), size = n(), replace = TRUE)) %>%
+    mutate(ANL01FL = ifelse(.data$TRTEMFL == "Y" & .data$ASTDTM <= .data$TRTEDTM + 2592000, "Y", ""))
+
 
   # Split metadata for AEs of special interest (AESI)
   l_AAG <- split(AAG, interaction(AAG$NAMVAR, AAG$SRCVAR, AAG$GRPTYPE, drop = TRUE)) # nolint
@@ -242,12 +247,12 @@ radae <- function(ADSL, # nolint
     }
 
     d_adag <- d_adag[-which(names(d_adag) %in% c("NAMVAR", "SRCVAR", "GRPTYPE"))] # nolint
-    d_new <- dplyr::left_join(x = d_adae, y = d_adag, by = intersect(names(d_adae), names(d_adag)))
-    d_new[, dplyr::setdiff(names(d_new), names(d_adae)), drop = FALSE]
+    d_new <- left_join(x = d_adae, y = d_adag, by = intersect(names(d_adae), names(d_adag)))
+    d_new[, setdiff(names(d_new), names(d_adae)), drop = FALSE]
 
   }, ADAE)
 
-  ADAE <- dplyr::bind_cols(ADAE, l_AESI) # nolint
+  ADAE <- bind_cols(ADAE, l_AESI) # nolint
 
   # apply metadata
   ADAE <- apply_metadata(ADAE, "metadata/ADAE.yml") # nolint